+++ frontend/src/component/MenuBar.jsx
@@ -2,9 +2,11 @@ import React, { Component } from "react";
 import {
     MinusOutlined,
     FullscreenOutlined,
+    FullscreenExitOutlined,
     PoweroffOutlined
 } from "@ant-design/icons";
 import Logo from "./Logo";
+import store from "../store";
 const ipcRenderer = window.ipcRenderer;
 
 // import Service from "../Service";
@@ -18,9 +20,19 @@ const ipcRenderer = window.ipcRenderer;
 class MenuBar extends Component {
     constructor(props) {
         super(props);
+        this.state = store.getState();
+        this.storageChange = this.storageChange.bind(this);
         this.minWin = this.minWin.bind(this);
         this.maxWin = this.maxWin.bind(this);
         this.closeWin = this.closeWin.bind(this);
+    }
+
+    storageChange() {
+        this.setState(store.getState());
+    }
+
+    componentDidMount() {
+        store.subscribe(this.storageChange);
     }
 
     minWin() {
@@ -29,6 +41,12 @@ class MenuBar extends Component {
 
     maxWin() {
         ipcRenderer.send("window-full");
+        const lastState = this.state.fullScreened;
+        this.setState(() => {
+            return {
+                fullScreened: lastState ? false : true
+            };
+        });
     }
 
     closeWin() {
@@ -46,15 +64,25 @@ class MenuBar extends Component {
                         innerColor={"base-logo"}
                     />
                 </div>
-                <p className={`${"menubar-name"} ${"white-color"}`}>{"页页"}</p>
+                <p className={`${"menubar-name"} ${"white-color"}`}>
+                    {this.state.text.softName}
+                </p>
                 <MinusOutlined
                     className={`${"menubar-icon"} ${"menuIcon"}`}
                     onClick={this.minWin}
                 />
-                <FullscreenOutlined
-                    className={`${"menubar-icon"} ${"menuIcon"}`}
-                    onClick={this.maxWin}
-                />
+                {this.state.fullScreened ? (
+                    <FullscreenExitOutlined
+                        className={`${"menubar-icon"} ${"menuIcon"}`}
+                        onClick={this.maxWin}
+                    />
+                ) : (
+                    <FullscreenOutlined
+                        className={`${"menubar-icon"} ${"menuIcon"}`}
+                        onClick={this.maxWin}
+                    />
+                )}
+
                 <PoweroffOutlined
                     className={`${"menubar-icon"} ${"menuIcon"}`}
                     onClick={this.closeWin}