+++ frontend/src/local/en_US.jsx
@@ -1,0 +1,17 @@
+const en_US = {
+    softName: "Ye Ye",
+    camera: "CAMERA",
+    upload: "UPLOAD",
+    history: "HISTORY",
+    metronome: "BEAT",
+    timer: "ENABLE",
+    speed: "SPEED",
+    speedInfo: "Move to the left, face recognition is faster",
+    language: "LANGUAGE",
+    theme: "THEME",
+    about: "ABOUT",
+    collapse: "COLLAPSE",
+    delete: "DELETE"
+};
+
+export default en_US;