+++ frontend/src/local/zh_CN.jsx
@@ -1,0 +1,17 @@
+const zh_CN = {
+    softName: "页页",
+    camera: "摇头翻页",
+    upload: "上传文件",
+    history: "历史记录",
+    metronome: "节拍器",
+    timer: "自动翻页",
+    speed: "翻页调速",
+    speedInfo: "往左移动，人脸识别更快哦",
+    language: "中英切换",
+    theme: "主题切换",
+    about: "关于我们",
+    collapse: "收起",
+    delete: "删除"
+};
+
+export default zh_CN;