+++ frontend/src/store/reducer.jsx
@@ -1,4 +1,8 @@
 import * as Type from "./actionTypes";
+
+const en = require("../local/en_US");
+const zh = require("../local/zh_CN");
+
 const defaultState = {
     // open or close camera
     cameraOpened: false,
@@ -18,7 +22,12 @@ const defaultState = {
     // but History.Drawer don't support less, so...
     colors: ["#FFFFFF", "#373B43", "#72767E", "#535860", "#1D212A", "#141D2F"],
     // menu stretch
-    toolCollapsed: true
+    toolCollapsed: true,
+    // Chinese / English
+    useZh: true,
+    text: zh.default,
+    // full/window screen state
+    fullScreened: true
 };
 
 // action = {type: string, value: any}
@@ -72,6 +81,17 @@ export default (state = defaultState, action = { type: "noneType" }) => {
         case Type.SWITCH_TOOL:
             newState = JSON.parse(JSON.stringify(state));
             newState.toolCollapsed = action.value;
+            return newState;
+        // language state
+        case Type.SWITCH_LANGUAGE:
+            newState = JSON.parse(JSON.stringify(state));
+            newState.useZh = action.value;
+            newState.text = action.value ? zh.default : en.default;
+            return newState;
+        // full/window screen state
+        case Type.FULL_SCREEN:
+            newState = JSON.parse(JSON.stringify(state));
+            newState.fullScreened = action.value;
             return newState;
         default:
             return state;